+++ Quick.cpp
@@ -1,13 +1,30 @@
 ﻿#include "sort.h"
 
-void InsertionSort(int a[], int n) {
-	int i = 0, j = 0, key = 0;
-	for (int i = 1; i < n; i++) {
-		key = a[i];
-		j = i - 1;
-		while (j >= 0 && a[j] > key) {
-			a[j + 1] = a[j];
-			j = j - 1;
-		}a[j + 1] = key;
+int Partition(int a[], int left, int right) {
+
+	int pivot = a[right];
+	int i, j;
+	i = (left - 1);
+	for (j = left; j <= right - 1; j++){
+		if (a[j] < pivot){
+			i++;
+			swap(a[i], a[j]);
+		}
 	}
+	swap(a[i + 1], a[right]);
+	return (i + 1);
+}
+
+void QuickSortSub(int a[], int left, int right) {
+	if (left < right)
+	{
+		int pi = Partition(a, left, right);
+
+		QuickSortSub(a, left, pi - 1);  // Before pi
+		QuickSortSub(a, pi + 1, right); // After pi
+	}
+}
+
+void QuickSort(int a[], int n){
+	return QuickSortSub(a,0,n);
 }
